+++ gestao_residuos/src/main/java/br/com/fiap/gestao_residuos/dto/ResiduoExibicaoDTO.java
@@ -2,12 +2,14 @@ package br.com.fiap.gestao_residuos.dto;
 
 import br.com.fiap.gestao_residuos.model.Residuo;
 
+import java.math.BigDecimal;
+
 public record ResiduoExibicaoDTO(
 
         Long id,
         String nome,
         String tipo,
-        Double peso
+        BigDecimal peso
 ) {
     public ResiduoExibicaoDTO (Residuo residuo){
         this(
